Clean up and splitting
