+++ src/functionalTest/java/uk/gov/hmcts/reform/fact/functional/helpers/TestDataHelper.java
@@ -7,7 +7,9 @@ import io.restassured.response.Response;
 import uk.gov.hmcts.reform.fact.data.api.entities.Court;
 import uk.gov.hmcts.reform.fact.data.api.entities.CourtFacilities;
 import uk.gov.hmcts.reform.fact.data.api.entities.CourtLock;
+import uk.gov.hmcts.reform.fact.data.api.entities.ServiceCentre;
 import uk.gov.hmcts.reform.fact.data.api.entities.User;
+import uk.gov.hmcts.reform.fact.data.api.entities.types.CatchmentType;
 import uk.gov.hmcts.reform.fact.data.api.entities.types.Page;
 import uk.gov.hmcts.reform.fact.data.api.entities.types.UserRole;
 import uk.gov.hmcts.reform.fact.data.api.os.OsData;
@@ -160,6 +162,68 @@ public final class TestDataHelper {
         court.setOpenOnCath(openOnCath);
 
         final Response createResponse = http.doPost("/courts/v1", court);
+        assertThat(createResponse.statusCode()).isEqualTo(CREATED.value());
+
+        return UUID.fromString(createResponse.jsonPath().getString("id"));
+    }
+
+    /**
+     * Builds a test service centre with the given name.
+     *
+     * @param http the HTTP client
+     * @param serviceCentreName the name for the test service centre
+     * @return the service centre body
+     */
+    public static ServiceCentre buildServiceCentre(final HttpClient http, final String serviceCentreName) {
+        return buildServiceCentre(http, serviceCentreName, HttpClient.getAdminBearerToken());
+    }
+
+    /**
+     * Builds a test service centre with the given name.
+     *
+     * @param http the HTTP client
+     * @param serviceCentreName the name for the test service centre
+     * @param bearerToken the bearer token to use when fetching reference data
+     * @return the service centre body
+     */
+    public static ServiceCentre buildServiceCentre(final HttpClient http,
+                                                  final String serviceCentreName,
+                                                  final String bearerToken) {
+        final ServiceCentre serviceCentre = new ServiceCentre();
+        serviceCentre.setName(appendRandomSuffixToCourtName(serviceCentreName));
+        serviceCentre.setOpen(false);
+        serviceCentre.setServiceAreaIds(List.of(UUID.fromString(
+            http.doGet("/types/v1/service-areas", bearerToken).jsonPath().getString("[0].id")
+        )));
+        serviceCentre.setCatchmentType(CatchmentType.NATIONAL);
+        return serviceCentre;
+    }
+
+    /**
+     * Creates a test service centre with the given name.
+     *
+     * @param http the HTTP client
+     * @param serviceCentreName the name for the test service centre
+     * @return the created service centre UUID
+     */
+    public static UUID createServiceCentre(final HttpClient http, final String serviceCentreName) {
+        return createServiceCentre(http, serviceCentreName, HttpClient.getAdminBearerToken());
+    }
+
+    /**
+     * Creates a test service centre with the given name.
+     *
+     * @param http the HTTP client
+     * @param serviceCentreName the name for the test service centre
+     * @param bearerToken the bearer token to use when creating the service centre
+     * @return the created service centre UUID
+     */
+    public static UUID createServiceCentre(final HttpClient http,
+                                           final String serviceCentreName,
+                                           final String bearerToken) {
+        final ServiceCentre serviceCentre = buildServiceCentre(http, serviceCentreName, bearerToken);
+
+        final Response createResponse = http.doPost("/service-centres/v1", serviceCentre, bearerToken);
         assertThat(createResponse.statusCode()).isEqualTo(CREATED.value());
 
         return UUID.fromString(createResponse.jsonPath().getString("id"));
